Add files via upload
By reviewing the data, a few key points stick out immediately. For one, the outcomes reviewed in our data set may not be fully indicative of the data set as a whole since we are only looking at the Theatre category and Play subcategory. While, it may be a reasonable assumption to think that kickstarter projects in all categories may have similar success, failure and cancel rates, there is no way to be certain of this analysis unless we take a broader perspective on the data and conduct analysis on the holistic data set.

Secondly, per the Outcomes Based on Goals sheet, it can be derived that Plays with a goal in excess of $5,000.00 have a significantly higher chance of failing rather than being successful. This looks to be the case with each incremental increase in the funding goal target. The line chart also sheds light on the inverse relationship between the size of the funding target and it's success/failure rate. Projects with a lower funding target tend to have a growing chance of success, which eventually plateaus and projects with a higher funding target tend to have a decreasing chance of success, eventually bottoming out at a zero percent success rate.
 
Through review of the Outcomes Based on Launch Date data, the middle of the year looks like peak season both in terms of the number of projects that are submitted as well as those that are successful. Based on the 2 angles of analysis above, I would propose for those who are interested in starting Plays to propose projects on kickstarter with a funding goal of < $5000.00 and between the months of April - July to allow for the best chance of success.